requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pytest
httpx
pymongo
fastapi[standard]
mangum
transformers
torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu118